+++ src/side_area_panel/modules/dp_reorder/__init__.py
@@ -14,6 +14,3 @@
 #
 #  You should have received a copy of the GNU General Public License along with
 #  StatPrism.  If not, see <https://www.gnu.org/licenses/>.
-
-
-version = "1.2.6"